requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cheshire_cat_api==1.4.1
dataclasses-json==0.6.4
exceptiongroup==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain-community==0.0.19
langchain-core==0.1.22
langsmith==0.0.87
marshmallow==3.20.2
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pydantic==2.6.1
pydantic_core==2.16.2
pygame==2.5.2
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.0.7
websocket-client==1.7.0
yarl==1.9.4
